chore : (Comment) DDD layer에 맞는 모듈/디렉토리 설정
