+++ orury-client/src/main/java/org/orury/client/comment/application/response/CommentsWithCursorResponse.java
@@ -1,4 +1,4 @@
-package org.orury.client.comment.converter.response;
+package org.orury.client.comment.application.response;
 
 import org.orury.domain.global.constants.NumberConstants;
