+++ orury-client/src/main/java/org/orury/client/comment/interfaces/CommentFacade.java
@@ -1,0 +1,4 @@
+package org.orury.client.comment.interfaces;
+
+public class CommentFacade {
+}
