+++ orury-client/src/main/java/org/orury/client/user/controller/UserController.java
@@ -3,7 +3,7 @@ package org.orury.client.user.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.orury.client.comment.service.CommentService;
+import org.orury.client.global.WithCursorResponse;
 import org.orury.client.post.service.PostService;
 import org.orury.client.review.service.ReviewService;
 import org.orury.client.user.converter.message.UserMessage;
@@ -13,9 +13,9 @@ import org.orury.client.user.converter.response.MyPostResponse;
 import org.orury.client.user.converter.response.MyReviewResponse;
 import org.orury.client.user.converter.response.MypageResponse;
 import org.orury.client.user.service.UserService;
-import org.orury.client.global.WithCursorResponse;
 import org.orury.domain.base.converter.ApiResponse;
-import org.orury.domain.comment.dto.CommentDto;
+import org.orury.domain.comment.domain.CommentService;
+import org.orury.domain.comment.domain.dto.CommentDto;
 import org.orury.domain.global.constants.NumberConstants;
 import org.orury.domain.post.dto.PostDto;
 import org.orury.domain.review.dto.ReviewDto;
