+++ orury-client/src/main/java/org/orury/client/user/converter/response/MyCommentResponse.java
@@ -2,7 +2,7 @@ package org.orury.client.user.converter.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import org.orury.client.global.IdIdentifiable;
-import org.orury.domain.comment.dto.CommentDto;
+import org.orury.domain.comment.domain.dto.CommentDto;
 
 import java.time.LocalDateTime;
