+++ orury-client/src/main/java/org/orury/client/user/service/UserService.java
@@ -6,8 +6,8 @@ import org.orury.common.error.code.UserErrorCode;
 import org.orury.common.error.exception.BusinessException;
 import org.orury.common.util.ImageUrlConverter;
 import org.orury.common.util.S3Folder;
-import org.orury.domain.comment.db.repository.CommentLikeRepository;
-import org.orury.domain.comment.db.repository.CommentRepository;
+import org.orury.domain.comment.infrastructure.CommentLikeRepository;
+import org.orury.domain.comment.infrastructure.CommentRepository;
 import org.orury.domain.global.constants.NumberConstants;
 import org.orury.domain.global.domain.ImageUtils;
 import org.orury.domain.gym.domain.GymStore;
