+++ orury-client/src/test/java/org/orury/client/comment/service/CommentLikeServiceTest.java
@@ -2,12 +2,12 @@ package org.orury.client.comment.service;
 
 import org.orury.common.error.code.CommentErrorCode;
 import org.orury.common.error.exception.BusinessException;
-import org.orury.domain.comment.db.model.Comment;
-import org.orury.domain.comment.db.model.CommentLike;
-import org.orury.domain.comment.db.model.CommentLikePK;
-import org.orury.domain.comment.db.repository.CommentLikeRepository;
-import org.orury.domain.comment.db.repository.CommentRepository;
-import org.orury.domain.comment.dto.CommentLikeDto;
+import org.orury.domain.comment.domain.entity.Comment;
+import org.orury.domain.comment.domain.entity.CommentLike;
+import org.orury.domain.comment.domain.entity.CommentLikePK;
+import org.orury.domain.comment.infrastructure.CommentLikeRepository;
+import org.orury.domain.comment.infrastructure.CommentRepository;
+import org.orury.domain.comment.domain.dto.CommentLikeDto;
 import org.orury.domain.global.constants.NumberConstants;
 import org.orury.domain.post.db.model.Post;
 import org.orury.domain.user.db.model.User;
