+++ orury-client/src/test/java/org/orury/client/comment/service/CommentServiceTest.java
@@ -1,26 +1,27 @@
 package org.orury.client.comment.service;
 
-import org.orury.common.error.code.CommentErrorCode;
-import org.orury.common.error.exception.BusinessException;
-import org.orury.domain.global.domain.ImageUtils;
-import org.orury.domain.comment.db.model.Comment;
-import org.orury.domain.comment.db.model.CommentLike;
-import org.orury.domain.comment.db.model.CommentLikePK;
-import org.orury.domain.comment.db.repository.CommentLikeRepository;
-import org.orury.domain.comment.db.repository.CommentRepository;
-import org.orury.domain.comment.dto.CommentDto;
-import org.orury.domain.comment.dto.CommentLikeDto;
-import org.orury.domain.global.constants.NumberConstants;
-import org.orury.domain.post.db.model.Post;
-import org.orury.domain.post.db.repository.PostRepository;
-import org.orury.domain.post.dto.PostDto;
-import org.orury.domain.user.db.model.User;
-import org.orury.domain.user.dto.UserDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.orury.common.error.code.CommentErrorCode;
+import org.orury.common.error.exception.BusinessException;
+import org.orury.domain.comment.domain.CommentService;
+import org.orury.domain.comment.domain.dto.CommentDto;
+import org.orury.domain.comment.domain.dto.CommentLikeDto;
+import org.orury.domain.comment.domain.entity.Comment;
+import org.orury.domain.comment.domain.entity.CommentLike;
+import org.orury.domain.comment.domain.entity.CommentLikePK;
+import org.orury.domain.comment.infrastructure.CommentLikeRepository;
+import org.orury.domain.comment.infrastructure.CommentRepository;
+import org.orury.domain.global.constants.NumberConstants;
+import org.orury.domain.global.domain.ImageUtils;
+import org.orury.domain.post.db.model.Post;
+import org.orury.domain.post.db.repository.PostRepository;
+import org.orury.domain.post.dto.PostDto;
+import org.orury.domain.user.db.model.User;
+import org.orury.domain.user.dto.UserDto;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.test.context.ActiveProfiles;
