--- orury-domain/src/main/java/org/orury/domain/comment/db/repository/Test.java
@@ -1,4 +1,0 @@
-package org.orury.domain.comment.db.repository;
-
-public class Test {
-}
