+++ orury-domain/src/main/java/org/orury/domain/comment/domain/entity/CommentLike.java
@@ -1,4 +1,4 @@
-package org.orury.domain.comment.db.model;
+package org.orury.domain.comment.domain.entity;
 
 import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
