+++ orury-domain/src/main/java/org/orury/domain/comment/infrastructure/CommentLikeRepository.java
@@ -1,7 +1,7 @@
-package org.orury.domain.comment.db.repository;
+package org.orury.domain.comment.infrastructure;
 
-import org.orury.domain.comment.db.model.CommentLike;
-import org.orury.domain.comment.db.model.CommentLikePK;
+import org.orury.domain.comment.domain.entity.CommentLike;
+import org.orury.domain.comment.domain.entity.CommentLikePK;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
