+++ orury-domain/src/main/java/org/orury/domain/comment/infrastructure/CommentRepository.java
@@ -1,6 +1,6 @@
-package org.orury.domain.comment.db.repository;
+package org.orury.domain.comment.infrastructure;
 
-import org.orury.domain.comment.db.model.Comment;
+import org.orury.domain.comment.domain.entity.Comment;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
