+++ orury-domain/src/main/java/org/orury/domain/global/constants/Constants.java
@@ -19,6 +19,7 @@ public enum Constants {
 
     // deleted values
     DELETED_USER_NICKNAME("탈퇴 유저"),
+    DELETED_COMMENT_CONTENT("삭제된 댓글입니다."),
 
     // 암장 조회 시, kakaoId 기반으로 kakaoMapLink 만들기 위한 baseUrl
     KAKAO_MAP_BASE_URL("https://map.kakao.com/?itemId=");
